+++ .github/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), and this project *somewhat* adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2020-06-07
+### Changed
+ - Hearthstone Metadata endpoint files are now automatically generated from a CSV
+   - This also fixed incorrect API paths with most metadata
+
 ## [1.2.0] - 2020-06-05
 ### Added
  - Added support for the following Hearthstone Game Data endpoint category :